+++ js/main.js
@@ -1,2 +1,12 @@
-import './data.js';
-import './util.js';
+//import {getRandomIntInclusive, getRandomFloat, advertisementsArrayGenerator} from '.data.js';
+import {createAuthor, createLocation,createOffer} from './data.js';
+import {advertisementsArrayGenerator} from './util.js';
+/*
+getRandomIntInclusive(1, 10);
+getRandomFloat(1, 10, 2);
+advertisementsArrayGenerator();
+*/
+createAuthor();
+createLocation();
+createOffer();
+advertisementsArrayGenerator();
